# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API de paiement mobile OM Pay pour la gestion bancaire",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/ompay/transactions/stats": {
            "get": {
                "tags": [
                    "Transactions OM Pay"
                ],
                "summary": "Récupérer les statistiques des transactions",
                "description": "Retourne les statistiques des transactions OM Pay de l'utilisateur",
                "operationId": "f5aba456c3a1da4a931cc526aaeec379",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_transactions": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 12500
                                                },
                                                "successful_transactions": {
                                                    "type": "integer",
                                                    "example": 23
                                                },
                                                "failed_transactions": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "pending_transactions": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "average_transaction": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                },
                                                "monthly_stats": {
                                                    "properties": {
                                                        "current_month": {
                                                            "properties": {
                                                                "count": {
                                                                    "type": "integer",
                                                                    "example": 5
                                                                },
                                                                "amount": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 2500
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu après authentification OM Pay. Format: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}